Abstracting code for sources
