+++ components/openai/sources/common.mjs
@@ -1,0 +1,34 @@
+import { DEFAULT_POLLING_SOURCE_TIMER_INTERVAL } from "@pipedream/platform";
+import openai from "../../openai.app.mjs";
+
+export default {
+  props: {
+    openai,
+    db: "$.service.db",
+    timer: {
+      type: "$.interface.timer",
+      default: {
+        intervalSeconds: DEFAULT_POLLING_SOURCE_TIMER_INTERVAL,
+      },
+    },
+  },
+  methods: {
+    _getSavedItems() {
+      return this.db.get("savedItems") ?? [];
+    },
+    _setSavedItems(value) {
+      this.db.set("savedItems", value);
+    },
+    async getAndProcessItems() {
+      throw new Error("No item fetching implemented!");
+    },
+  },
+  hooks: {
+    async deploy() {
+      await this.getAndProcessItems();
+    },
+  },
+  async run() {
+    await this.getAndProcessItems();
+  },
+};
